* Day 10 Solved (amendment): Realized that the directional map struct really   should be using a signed variable type because it contains negative numbers.   Luckily the compiler seemed to be smart enough to work as it was but I'm   updating it for correctness sake.

+++ day_10/main.c
@@ -28,8 +28,8 @@ struct visisted
 
 const struct dirMap
 {
-	const size_t delta_x;
-	const size_t delta_y;
+	const int delta_x;
+	const int delta_y;
 } dir_map[] = 
 {
 	{0, -1}, /* NORTH */
